Move TestData to its original place
