+++ metricbeat/module/etcd/leader/leader_integration_test.go
@@ -45,6 +45,21 @@ func TestFetch(t *testing.T) {
 	t.Logf("%s/%s event: %+v", f.Module().Name(), f.Name(), events[0])
 }
 
+func TestData(t *testing.T) {
+	compose.EnsureUp(t, "etcd")
+
+	f := mbtest.NewReportingMetricSetV2(t, getConfig())
+	events, errs := mbtest.ReportingFetchV2(f)
+	if len(errs) > 0 {
+		t.Fatalf("Expected 0 error, had %d. %v\n", len(errs), errs)
+	}
+	assert.NotEmpty(t, events)
+
+	if err := mbtest.WriteEventsReporterV2(f, t, ""); err != nil {
+		t.Fatal("write", err)
+	}
+}
+
 func getConfig() map[string]interface{} {
 	return map[string]interface{}{
 		"module":     "etcd",
